Extract notes into NotesMapper
