+++ app/services/cocina/from_fedora/descriptive.rb
@@ -18,7 +18,7 @@ module Cocina
 
       def props
         { title: [{ status: 'primary', value: TitleMapper.build(item) }] }.tap do |desc|
-          desc[:note] = notes if notes.present?
+          desc[:note] = NotesMapper.build(item)
           desc[:language] = language if language.present?
           desc[:contributor] = contributor if contributor.present?
           desc[:form] = form if form.present?
@@ -28,45 +28,6 @@ module Cocina
       private
 
       attr_reader :item
-
-      def notes
-        @notes ||= [].tap do |items|
-          items << original_url if original_url
-          items << abstract if abstract
-          items << statement_of_responsibility if statement_of_responsibility
-          items << thesis_statement if thesis_statement
-          additional_notes.each do |note|
-            items << { value: note.content }
-          end
-        end
-      end
-
-      def abstract
-        return if item.descMetadata.abstract.blank?
-
-        @abstract ||= { type: 'summary', value: item.descMetadata.abstract.first }
-      end
-
-      # TODO: Figure out how to encode displayLabel https://github.com/sul-dlss/dor-services-app/issues/849#issuecomment-635713964
-      def original_url
-        val = item.descMetadata.ng_xml.xpath('//mods:note[@type="system details"][@displayLabel="Original site"]', mods: DESC_METADATA_NS).first
-        { type: 'system details', value: val.content } if val
-      end
-
-      def statement_of_responsibility
-        val = item.descMetadata.ng_xml.xpath('//mods:note[@type="statement of responsibility"]', mods: DESC_METADATA_NS).first
-        { type: 'statement of responsibility', value: val.content } if val
-      end
-
-      def thesis_statement
-        val = item.descMetadata.ng_xml.xpath('//mods:note[@type="thesis"]', mods: DESC_METADATA_NS).first
-        { type: 'thesis', value: val.content } if val
-      end
-
-      # Returns any notes values that do not include a type attribute
-      def additional_notes
-        item.descMetadata.ng_xml.xpath('//mods:note[not(@type)]', mods: DESC_METADATA_NS)
-      end
 
       def language
         @language ||= [].tap do |langs|
